refactor: extract transfer timing and rate stats into transfer-stats module
